requirements.txt
chainlit==1.0.0
pydantic>=2.0.0,<2.10.0
google-genai>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
faiss-cpu>=1.9.0
pypdf==3.17.4
python-dotenv==1.0.0
